# CONNECT-HUB-Api-Gateway/cmd/docs/swagger.yaml
definitions:
...
  models.JobSeekerLogin:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  models.JobSeekerProfile:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone_number:
        type: string
    type: object
  models.JobSeekerSignUp:
    properties:
      confirm_password:
        maxLength: 20
        minLength: 6
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
      phone_number:
        type: string
    required:
    - confirm_password
    - email
    - password
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: ConnectHub is a professional networking platform.
  title: Go + Gin professional networking platform API Connect Hub
  version: 1.0.0
paths:
...
  /jobseeker/login:
    post:
      consumes:
      - application/json
      description: Allows a job seeker to log in by providing necessary credentials.
      parameters:
      - description: Job Seeker Login Data
        in: body
        name: jobseekerData
        required: true
        schema:
          $ref: '#/definitions/models.JobSeekerLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Jobseeker authenticated successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Incorrect data format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Job Seeker Login
      tags:
      - Jobseeker Authentication Management
  /jobseeker/policies:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all policies.
      produces:
      - application/json
      responses:
        "200":
          description: Policies retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to retrieve policies
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get All Policies
      tags:
      - Jobseeker Policies Management
  /jobseeker/policy:
    get:
      consumes:
      - application/json
      description: Retrieves a specific policy by its ID.
      parameters:
      - description: Policy ID
        in: query
        name: policy_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Policy retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to retrieve policy
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get One Policy
      tags:
      - Jobseeker Policies Management
...
  /jobseeker/profile:
    get:
      consumes:
      - application/json
      description: Retrieves the profile of the logged-in job seeker.
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Incorrect data format or failed to retrieve profile
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get Job Seeker Profile
      tags:
      - Jobseeker Profile Management
    patch:
      consumes:
      - application/json
      description: Allows a job seeker to edit their profile by providing necessary
        information.
      parameters:
      - description: Job Seeker Profile Data
        in: body
        name: jobseekerData
        required: true
        schema:
          $ref: '#/definitions/models.JobSeekerProfile'
      produces:
      - application/json
      responses:
        "200":
          description: Profile edited successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Incorrect data format or failed to edit profile
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Edit Job Seeker Profile
      tags:
      - Jobseeker Profile Management
  /jobseeker/signup:
    post:
      consumes:
      - application/json
      description: Allows a job seeker to sign up by providing necessary information.
      parameters:
      - description: Job Seeker Sign Up Data
        in: body
        name: jobseekerData
        required: true
        schema:
          $ref: '#/definitions/models.JobSeekerSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: Jobseeker signup successful
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Incorrect data format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Job Seeker Signup
      tags:
      - Jobseeker Authentication Management
...
  /recruiter/video-call/key:
    get:
      consumes:
      - application/json
      description: Generate a video call key and private link for interview.
      parameters:
      - description: 'User ID of the jobseeker '
        in: query
        name: user
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved video call key and private link
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid user ID format or role mismatch
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Only recruiters can access this endpoint'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal Server Error: Failed to generate video call key'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Generate Video Call Key
      tags:
      - Recruiter Videocall Management
securityDefinitions:
  BearerTokenAuth:
    description: 'For accessing the API, you need to include "Jobseeker","Admin" or
      "Recruiter" before your token in the Authorization header. Example: "Authorization:
      Admin <your_token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"